my_selenium_projrct/requirements.txt
allure-pytest==2.8.34
allure-python-commons==2.8.34
atomicwrites==1.4.0
attrs==20.2.0
basepage==1.7.3
beautifulsoup4==4.9.3
bs4==0.0.1
by==0.0.2
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
docopt==0.6.2
emailutils==0.0.1
future==0.18.2
idna==2.10
iniconfig==1.1.1
itchat==1.2.32
login==0.0.6
lxml==4.6.2
packaging==20.4
pipreqs==0.4.10
pluggy==0.13.1
py==1.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
pytest==6.1.1
pytest-assume==2.4.2
PyYAML==5.4.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
testreport==0.3
toml==0.10.1
urllib3==1.25.10
wxpy==0.3.9.8
xlrd==2.0.1
yarg==0.1.9
yattag==1.14.0
